requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary
gunicorn>=21.2
django-jazzmin>=2.6
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
cloudinary>=1.36
python-dotenv>=1.0
Pillow